Delete a task will delete related assignments

+++ jobs/task_assignments/delete.js
@@ -1,6 +1,4 @@
 const models = require('../../models');
-let Project = models.projects;
-let ProjectAssignment = models.project_assignments;
 let User = models.users;
 let Task = models.tasks;
 let TaskAssignment = models.task_assignments;
@@ -19,20 +17,61 @@ module.exports = function(connection, done) {
         // LOG
         console.log(" [%s]: %s", msg.fields.routingKey, msg.content.toString());
         let json = JSON.parse(msg.content.toString());
-        TaskAssignment.destroy({
+        Task.find({
           where: {
-            uuid: json.uuid
+            uuid: json.taskUuid
           }
-        }).then(function (assignment) {
-          ch.sendToQueue(msg.properties.replyTo,
-            new Buffer.from(JSON.stringify(assignment)),
-            { correlationId: msg.properties.correlationId });
-          connection.createChannel(function(error, channel) {
-            channel.assertExchange(ex, 'topic');
-            channel.publish(ex, queue + '.reply', new Buffer(msg.content.toString()));
-          });
-          ch.ack(msg);
+        }).then(function (task) {
+          if (task != undefined) {
+            User.find({
+              where: {
+                email: json.email
+              }
+            }).then(function (user) {
+              if (user != undefined || json.email == undefined) {
+                args = { taskId: task.id }
+                if (json.email != undefined) {
+                  args.userId = user.id;
+                }
+                TaskAssignment.destroy({
+                  where: args
+                }).then(function (assignment) {
+                  ch.sendToQueue(msg.properties.replyTo,
+                    new Buffer.from(JSON.stringify({ status: 'removed' })),
+                    { correlationId: msg.properties.correlationId });
+                  connection.createChannel(function(error, channel) {
+                    channel.assertExchange(ex, 'topic');
+                    channel.publish(ex, queue + '.reply', new Buffer(msg.content.toString()));
+                  });
+                  ch.ack(msg);
+                }).catch(function (error) {
+                  console.log(error);
+                  ch.sendToQueue(msg.properties.replyTo,
+                    new Buffer(error.toString()),
+                    { correlationId: msg.properties.correlationId });
+                  ch.ack(msg);
+                });
+              } else {
+                ch.sendToQueue(msg.properties.replyTo,
+                  new Buffer("Unknown task"),
+                  { correlationId: msg.properties.correlationId });
+                ch.ack(msg);
+              }
+            }).catch(function (error) {
+              console.log(error);
+              ch.sendToQueue(msg.properties.replyTo,
+                new Buffer(error.toString()),
+                { correlationId: msg.properties.correlationId });
+              ch.ack(msg);
+            });
+          } else {
+            ch.sendToQueue(msg.properties.replyTo,
+              new Buffer("Unknown user"),
+              { correlationId: msg.properties.correlationId });
+            ch.ack(msg);
+          }
         }).catch(function (error) {
+          console.log(error);
           ch.sendToQueue(msg.properties.replyTo,
             new Buffer(error.toString()),
             { correlationId: msg.properties.correlationId });
